chore: #91 move txn getters separate file
